+++ src/Validate.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * SimpleComplex PHP Validate
+ * @link      https://github.com/simplecomplex/php-validate
+ * @copyright Copyright (c) 2020 Jacob Friis Mathiasen
+ * @license   https://github.com/simplecomplex/php-validate/blob/master/LICENSE (MIT License)
+ */
+declare(strict_types=1);
+
+namespace SimpleComplex\Validate;
+
+/**
+ * @deprecated  Use Validator instead.
+ *
+ * @package SimpleComplex\Validate
+ */
+class Validate extends Validator
+{
+}